modifications to allow grid templates to work with odd or even widths
Former-commit-id: 1ce72d44a74604bbe0778aa976d0a348b36ceddb

+++ met/src/basic/vx_util/GridTemplate.h
@@ -116,7 +116,7 @@ class GridTemplate
 
   virtual const char* getClassName(void) const = 0;
 
-  virtual double getWidth() const = 0;
+  virtual int getWidth() const = 0;
   
  protected:
